feat: Add zsa udev rule
This is for my Moonlander Keyboard from zsa that I need to flash from
time to time.

Using Chrome, this allows the browser app to send low level commands to
the keyboard so when I make changes to my keyboard layout, I can load
them onto the hardware of my computer.